blog: First draft
Introduce an initial build system to separate the build proces of the website.

Documentation and key material is still stored in the libcamera repository,
but any blog entries are considered private and maintiained here.

site/conf.py and site/index.rst are duplicated directly from our main
repository and will be updated locally.

The other components are simply symlinks to the main repository, which
is cloned at 'libcamera/' by the makefile.

This clone could be replaced by a symlink to an external repository.

Signed-off-by: Kieran Bingham <kieran@bingham.xyz>